+++ src/main/java/com/rezdy/lunch/dto/Ingredient.java
@@ -1,4 +1,4 @@
-package com.rezdy.lunch.service;
+package com.rezdy.lunch.dto;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
