+++ src/main/java/com/rezdy/lunch/repository/RecipeRepository.java
@@ -1,6 +1,6 @@
 package com.rezdy.lunch.repository;
 
-import com.rezdy.lunch.service.Recipe;
+import com.rezdy.lunch.dto.Recipe;
 
 import java.time.LocalDate;
 import java.util.List;
