+++ src/main/java/com/rezdy/lunch/repository/RecipeRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.rezdy.lunch.repository;
 
-import com.rezdy.lunch.service.Recipe;
+import com.rezdy.lunch.dto.Recipe;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
